Move LowerT & Dist into Control.Monad.Trans.Lower.
